cleanup for gql only
