--- src/resolvers/mutations/index.js
@@ -1,5 +1,0 @@
-const createUser = require('./createUser');
-const updateUser = require('./updateUser');
-const deleteUser = require('./deleteUser');
-
-module.exports = { createUser, updateUser, deleteUser };
