--- src/resolvers/queries/index.js
@@ -1,4 +1,0 @@
-const users = require('./users.js');
-const user = require('./user.js');
-
-module.exports = { users, user };
