--- src/resolvers/queries/user.js
@@ -1,5 +1,0 @@
-const { User } = require('../../models/User');
-
-module.exports = async (_, { id }, { models }) => {
-  return await User.findOne({ _id: id });
-};
