--- src/resolvers/queries/users.js
@@ -1,5 +1,0 @@
-const { User } = require('../../models/User');
-
-module.exports = async (_, {}, { models }) => {
-  return await User.find();
-};
